added all apis

+++ src/Module/userModule/userModule.controller.ts
@@ -31,33 +31,37 @@ export class UserController {
     @Res({ passthrough: true }) res: Response,
     @Query('page') page: number = 1,
   ) {
-    // if (authUser.role === 'regular' || authUser.role === 'manager') {
-    //   return res
-    //     .status(HttpStatus.BAD_REQUEST)
-    //     .send({ msg: 'Regular/Manager user cannot perform any function' });
-    // }
+    try {
+      if (authUser.role === 'regular' || authUser.role === 'manager') {
+        return res
+          .status(HttpStatus.BAD_REQUEST)
+          .send({ msg: 'Regular/Manager user cannot perform any function' });
+      }
 
-    const schema = Joi.object()
-      .options({ abortEarly: false })
-      .keys({
-        page: Joi.number().min(1).required(),
-      })
-      .unknown();
+      const schema = Joi.object()
+        .options({ abortEarly: false })
+        .keys({
+          page: Joi.number().min(1).required(),
+        })
+        .unknown();
 
-    const valid = schema.validate({
-      page,
-    });
-    console.log(valid);
+      const valid = schema.validate({
+        page,
+      });
+      console.log(valid);
 
-    if (valid.error) {
-      return res.status(HttpStatus.BAD_REQUEST).send({ error: valid.error });
+      if (valid.error) {
+        return res.status(HttpStatus.BAD_REQUEST).send({ error: valid.error });
+      }
+
+      const Users = await this.userService.getAllUsers(page);
+      if ('msg' in Users) {
+        return Users;
+      }
+      return res.status(HttpStatus.BAD_REQUEST).send(Users);
+    } catch (e) {
+      return res.status(HttpStatus.BAD_REQUEST).send(e.message);
     }
-
-    const Users = await this.userService.getAllUsers(page);
-    if ('msg' in Users) {
-      return Users;
-    }
-    return res.status(HttpStatus.BAD_REQUEST).send(Users);
   }
 
   @Delete('deleteUser/:id')
@@ -66,16 +70,20 @@ export class UserController {
     @Param('id', new ParseIntPipe()) id,
     @Auth() authUser: IAuth,
   ) {
-    // if (authUser.role === 'regular') {
-    //   return res
-    //     .status(HttpStatus.BAD_REQUEST)
-    //     .send({ msg: 'Regular user cannot delete user' });
-    // }
-    const user = await this.userService.deleteUser(id);
-    if ('msg' in user) {
-      return user;
+    try {
+      if (authUser.role === 'regular' || authUser.role === 'manager') {
+        return res
+          .status(HttpStatus.BAD_REQUEST)
+          .send({ msg: 'Regular/Manager user cannot delete user' });
+      }
+      const user = await this.userService.deleteUser(id);
+      if ('msg' in user) {
+        return user;
+      }
+      return res.status(HttpStatus.BAD_REQUEST).send(user);
+    } catch (e) {
+      return res.status(HttpStatus.BAD_REQUEST).send(e.message);
     }
-    return res.status(HttpStatus.BAD_REQUEST).send(user);
   }
 
   @Post('createUser')
@@ -86,41 +94,45 @@ export class UserController {
     @Body('userId') userId: string,
     @Body('password') password: string,
   ) {
-    // if (authUser.role === 'regular') {
-    //   return res
-    //     .status(HttpStatus.BAD_REQUEST)
-    //     .send({ msg: 'Regular user cannot add user' });
-    // }
-    const schema = Joi.object()
-      .options({ abortEarly: false })
-      .keys({
-        role: Joi.string().valid('regular', 'manager', 'admin').required(),
-        userId: Joi.string().trim().min(7).max(24).required(),
-        password: Joi.string().trim().min(5).max(24).required(),
-      })
-      .unknown();
+    try {
+      if (authUser.role === 'regular' || authUser.role === 'manager') {
+        return res
+          .status(HttpStatus.BAD_REQUEST)
+          .send({ msg: 'Regular user cannot add user' });
+      }
+      const schema = Joi.object()
+        .options({ abortEarly: false })
+        .keys({
+          role: Joi.string().valid('regular', 'manager', 'admin').required(),
+          userId: Joi.string().trim().min(7).max(24).required(),
+          password: Joi.string().trim().min(5).max(24).required(),
+        })
+        .unknown();
 
-    const valid = schema.validate({
-      role: role,
-      userId: userId,
-      password: password,
-    });
+      const valid = schema.validate({
+        role: role,
+        userId: userId,
+        password: password,
+      });
 
-    if (valid.error) {
-      return res
-        .status(HttpStatus.BAD_REQUEST)
-        .send({ error: valid.error.message });
+      if (valid.error) {
+        return res
+          .status(HttpStatus.BAD_REQUEST)
+          .send({ error: valid.error.message });
+      }
+
+      const registeredBike = await this.userService.createUser({
+        role,
+        userId: userId.trim().toLowerCase(),
+        password: password.trim(),
+      });
+      if ('msg' in registeredBike) {
+        return registeredBike;
+      }
+      return res.status(HttpStatus.BAD_REQUEST).send(registeredBike);
+    } catch (e) {
+      return res.status(HttpStatus.BAD_REQUEST).send(e.message);
     }
-
-    const registeredBike = await this.userService.createUser({
-      role,
-      userId: userId.trim().toLowerCase(),
-      password: password.trim(),
-    });
-    if ('msg' in registeredBike) {
-      return registeredBike;
-    }
-    return res.status(HttpStatus.BAD_REQUEST).send(registeredBike);
   }
 
   @Patch('updateUser/:id')
@@ -131,38 +143,42 @@ export class UserController {
     @Body('userId') userId: string,
     @Auth() authUser: IAuth,
   ) {
-    // if (authUser.role === 'regular') {
-    //   return res
-    //     .status(HttpStatus.BAD_REQUEST)
-    //     .send({ msg: 'Regular user cannot update user' });
-    // }
-    const schema = Joi.object()
-      .options({ abortEarly: false })
-      .keys({
-        role: Joi.string().valid('regular', 'manager', 'admin').optional(),
-        userId: Joi.string().trim().min(7).max(24).optional(),
-      })
-      .unknown();
+    try {
+      if (authUser.role === 'regular' || authUser.role === 'manager') {
+        return res
+          .status(HttpStatus.BAD_REQUEST)
+          .send({ msg: 'Regular/Manager user cannot update user' });
+      }
+      const schema = Joi.object()
+        .options({ abortEarly: false })
+        .keys({
+          role: Joi.string().valid('regular', 'manager', 'admin').optional(),
+          userId: Joi.string().trim().min(7).max(24).optional(),
+        })
+        .unknown();
 
-    const valid = schema.validate({
-      role: role,
-      userId: userId,
-    });
+      const valid = schema.validate({
+        role: role,
+        userId: userId,
+      });
 
-    if (valid.error) {
-      return res
-        .status(HttpStatus.BAD_REQUEST)
-        .send({ error: valid.error.message });
+      if (valid.error) {
+        return res
+          .status(HttpStatus.BAD_REQUEST)
+          .send({ error: valid.error.message });
+      }
+
+      const users = await this.userService.updateUserForSpecificId({
+        id,
+        role,
+        userId,
+      });
+      if ('msg' in users) {
+        return users;
+      }
+      return res.status(HttpStatus.BAD_REQUEST).send(users);
+    } catch (e) {
+      return res.status(HttpStatus.BAD_REQUEST).send(e.message);
     }
-
-    const users = await this.userService.updateUserForSpecificId({
-      id,
-      role,
-      userId,
-    });
-    if ('msg' in users) {
-      return users;
-    }
-    return res.status(HttpStatus.BAD_REQUEST).send(users);
   }
 }
